Update submission progress check to work as a validation function, and related changes. (#1817)
