+++ backend/audit/cross_validation/__init__.py
@@ -59,9 +59,13 @@ from .additional_ueis import additional_ueis
 from .auditee_ueis_match import auditee_ueis_match
 from .audit_findings import audit_findings
 from .sac_validation_shape import sac_validation_shape  # noqa: F401
+from .submission_progress_check import submission_progress_check
+from .tribal_data_sharing_consent import tribal_data_sharing_consent
 
 functions = [
     audit_findings,
     auditee_ueis_match,
     additional_ueis,
+    submission_progress_check,
+    tribal_data_sharing_consent,
 ]
