+++ backend/audit/cross_validation/auditee_ueis_match.py
@@ -1,7 +1,7 @@
 from .errors import err_auditee_ueis_match
 
 
-def auditee_ueis_match(sac_dict):
+def auditee_ueis_match(sac_dict, *_args, **_kwargs):
     """Checks that the auditee_uei values in each sheet are the same."""
     all_sections = sac_dict["sf_sac_sections"]
     sections = filter(None, all_sections.values())
