+++ backend/audit/templates/audit/submission_checklist/submission-checklist.html
@@ -290,19 +290,13 @@
                                 <p class="text-bold text-underline margin-0">Pre-submission validation (Complete)</p>
                                 <p class="margin-0">Before submitting your audit, use our validation tool to check your workbooks for errors.</p>
                             </div>
-                        {% elif complete %}
+                        {% else %}
                             {% include './icon-list-icon.html' with completed=pre_submission_validation.completed %}
                             <div>
                                 <p class="text-bold margin-0">
                                     <a class="usa-link"
                                        href="{% url 'audit:CrossValidation' report_id %}">Pre-submission validation</a>
                                 </p>
-                                <p class="margin-0">Before submitting your audit, use our validation tool to check your workbooks for errors.</p>
-                            </div>
-                        {% else %}
-                            {% include './icon-list-icon.html' with completed=pre_submission_validation.completed %}
-                            <div>
-                                <p class="text-bold margin-0">Pre-submission validation</p>
                                 <p class="margin-0">Before submitting your audit, use our validation tool to check your workbooks for errors.</p>
                             </div>
                         {% endif %}
